requirements.txt
aiohttp==3.5.4
async-timeout==3.0.1
attrs==19.3.0
certifi==2019.9.11
chardet==3.0.4
discord.py==1.2.4
idna==2.8
idna-ssl==1.1.0
multidict==4.5.2
pyjq==2.4.0
python-dotenv==0.10.3
requests==2.22.0
typing-extensions==3.7.4
urllib3==1.25.6
websockets==6.0
yarl==1.3.0
